+++ src/style_parser.rs
@@ -1,5 +1,6 @@
-use std::{rc::Rc, cell::RefCell, convert::Infallible, collections::HashMap, hash::Hash};
+use std::{cell::RefCell, collections::HashMap, convert::Infallible, hash::Hash, rc::Rc};
 
+use indexmap::IndexMap;
 use lightningcss::{declaration::DeclarationBlock, properties::Property, rules::{keyframes::KeyframeSelector, CssRule}, stylesheet::{ParserOptions, PrinterOptions, StyleSheet}, traits::ToCss, visit_types, visitor::{Visit, VisitTypes, Visitor}};
 
 use crate::{style_propetries::{style_value_type::StyleValueType, unit::Platform}, utils::to_camel_case};
@@ -9,7 +10,7 @@ use super::parse_style_properties::parse_style_properties;
 pub type StyleValue = Vec<StyleValueType>;
 #[derive(Debug)]
 pub struct StyleData {
-  pub all_style: Rc<RefCell<HashMap<String, StyleValue>>>,
+  pub all_style: Rc<RefCell<IndexMap<String, StyleValue>>>,
 }
 
 pub struct KeyFramesData {
@@ -49,7 +50,6 @@ impl<'i> StyleVisitor<'i> {
 // 收集所有的样式到 all_style 中，以元祖的形式存在 (selector, vec[declaration1, declaration2, ...])
 impl<'i> Visitor<'i> for StyleVisitor<'i> {
   type Error = Infallible;
-  const TYPES: VisitTypes = visit_types!(RULES);
 
   fn visit_rule(&mut self, rule: &mut CssRule<'i>) -> Result<(), Self::Error> {
     match rule {
@@ -124,20 +124,23 @@ impl<'i> Visitor<'i> for StyleVisitor<'i> {
     }
     Ok(())
   }
+  
+  fn visit_types(&self) -> VisitTypes {
+       visit_types!(RULES)
+  }
+  
 }
 
 pub struct StyleParser<'i> {
   pub all_style: Rc<RefCell<Vec<(String, Vec<StyleDeclaration<'i>>)>>>,
   pub keyframes: Rc<RefCell<HashMap<String, Vec<KeyFrameItem>>>>,
-  pub platform: Platform
 }
 
 impl<'i> StyleParser<'i> {
-  pub fn new(platform:Platform) -> Self {
+  pub fn new(_: Platform) -> Self {
     StyleParser {
       all_style: Rc::new(RefCell::new(vec![])),
       keyframes: Rc::new(RefCell::new(HashMap::new())),
-      platform
     }
   }
 
@@ -185,7 +188,7 @@ impl<'i> StyleParser<'i> {
         ),
       )
     })
-    .collect::<HashMap<_, _>>();
+    .collect::<IndexMap<_, _>>();
 
     StyleData {
       all_style: Rc::new(RefCell::new(final_all_style)),