+++ src/style_propetries/background.rs
@@ -1,22 +1,18 @@
 use lightningcss::{
-    properties::{ background::Background as LNBackground, Property },
-    stylesheet::PrinterOptions,
-    targets::{ Features, Targets },
-    traits::ToCss,
-    values::color::CssColor,
+    printer::PrinterOptions, properties::{ background::Background as LNBackground, Property }, traits::ToCss, values::color::CssColor
 };
 use smallvec::SmallVec;
 
-use crate::{ generate_expr_lit_color, generate_expr_lit_str, generate_invalid_expr };
+use crate::generate_expr_lit_color;
 
 use super::{
-    background_image::{ parse_background_image_item, BackgroundImage },
+    background_image::{ parse_background_image_item, BackgroundImage, BackgroundImageKind },
     background_position::{ parse_background_position_item, BackgroundPosition },
     background_repeat::{ parse_background_repeat_item, BackgroundRepeat },
     background_size::{ parse_background_size_item, BackgroundSize },
     style_property_type::CSSPropertyType,
     traits::ToExpr,
-    unit::{ convert_color_keywords_to_hex, PropertyTuple },
+    unit::PropertyTuple,
 };
 
 fn parse_background(background: &SmallVec<[LNBackground<'_>; 1]>) -> Background {
@@ -36,6 +32,12 @@ fn parse_background(background: &SmallVec<[LNBackground<'_>; 1]>) -> Background 
         }
         background_repeat.push(parse_background_repeat_item(&item.repeat));
         if item.color != CssColor::default() {
+            background_color = Some(item.color.clone());
+        }
+        
+        if item.to_css_string(PrinterOptions::default()).unwrap() == "none" {
+            // 如果是none，就清空所有的值
+            background_image = vec![BackgroundImageKind::String("".to_string())];
             background_color = Some(item.color.clone());
         }
     }